chore: Move delete policy endpoint to endpoints section of app
